Purge all unused imports
