+++ src/test/java/seedu/pill/command/FindCommandTest.java
@@ -4,14 +4,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import seedu.pill.exceptions.PillException;
-import seedu.pill.exceptions.ExceptionMessages;
 import seedu.pill.util.Item;
 import seedu.pill.util.ItemMap;
 import seedu.pill.util.Storage;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
-import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
