+++ src/test/java/seedu/pill/command/TransactionsCommandTest.java
@@ -4,23 +4,20 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import seedu.pill.exceptions.PillException;
-import seedu.pill.util.ItemMap;
-import seedu.pill.util.Storage;
-import seedu.pill.util.Order;
-import seedu.pill.util.TransactionManager;
-import seedu.pill.util.Item;
+import seedu.pill.util.*;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
-import java.time.LocalDate;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * Unit tests for ViewOrdersCommand
+ * Unit tests for TransactionsCommand
  */
-public class ViewOrdersCommandTest {
+public class TransactionsCommandTest {
     private ItemMap itemMap;
     private Storage storage;
     private TransactionManager transactionManager;
@@ -39,33 +36,29 @@ public class ViewOrdersCommandTest {
     }
 
     @Test
-    public void execute_noOrders_printsEmptyList() throws PillException {
-        ViewOrdersCommand command = new ViewOrdersCommand(transactionManager);
+    public void execute_emptyTransactions_printsEmptyMessage() throws PillException {
+        TransactionsCommand command = new TransactionsCommand(transactionManager);
         command.execute(itemMap, storage);
-        String output = outputStream.toString().trim();
-        // Orders list is empty but no specific message is printed based on source code
-        assertTrue(output.isEmpty());
+        assertTrue(outputStream.toString().trim().contains("No transactions found"));
     }
 
     @Test
-    public void execute_withOrders_listsPendingOrders() throws PillException {
-        // Create an order with some items
-        ItemMap orderItems = new ItemMap();
-        orderItems.addItemSilent(new Item("Paracetamol", 10));
-        Order order = transactionManager.createOrder(Order.OrderType.PURCHASE, orderItems, "Test order");
+    public void execute_withTransactions_listsAllTransactions() throws PillException {
+        // Create a transaction by adding an incoming transaction
+        transactionManager.createTransaction("Paracetamol", 10,
+                Transaction.TransactionType.INCOMING, "Test transaction", null);
 
-        ViewOrdersCommand command = new ViewOrdersCommand(transactionManager);
+        TransactionsCommand command = new TransactionsCommand(transactionManager);
         command.execute(itemMap, storage);
 
         String output = outputStream.toString().trim();
         assertTrue(output.contains("Paracetamol"));
-        assertTrue(output.contains("10 in stock"));
-        assertTrue(output.contains("Test order"));
+        assertTrue(output.contains("10"));
     }
 
     @Test
     public void isExit_returnsAlwaysFalse() {
-        ViewOrdersCommand command = new ViewOrdersCommand(transactionManager);
+        TransactionsCommand command = new TransactionsCommand(transactionManager);
         assertFalse(command.isExit());
     }
