+++ src/test/java/seedu/pill/command/VisualizePriceCommandTest.java
@@ -3,7 +3,6 @@ package seedu.pill.command;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
-import seedu.pill.exceptions.ExceptionMessages;
 import seedu.pill.exceptions.PillException;
 import seedu.pill.util.ItemMap;
 import seedu.pill.util.Storage;
